requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
bcrypt
python-multipart
pydantic[email]
pydantic-settings
httpx
python-dotenv
alembic
google-generativeai
